# openapi.yaml
openapi: 3.0.3
info:
  title: J1Hub API
  version: 1.0.0
  description: |
    Stable API contract for J1Hub endpoints.

    ## Versioning plan
    - Canonical versioned base path: `/api/v1`.
    - New clients use `/api/v1/*`; unversioned routes remain available while `API_ENABLE_LEGACY_ROUTES=true`.
    - New clients should migrate to `/api/v1/*` immediately.

    ## Transition strategy
    1. **Now**: support both `/api/v1/*` and legacy endpoints.
    2. **Deprecation window**: communicate deprecation for legacy routes via release notes and docs.
    3. **Removal**: disable legacy aliases by configuration after migration completion.
servers:
  - url: /
    description: Legacy (transition period)
  - url: /api/v1
    description: Canonical versioned API
security:
  - bearerAuth: []
paths:
...
  /verify/upload:
    post:
      summary: Upload verification document
      operationId: uploadVerificationDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [document, waiver]
              properties:
                document:
                  type: string
                  format: binary
                waiver:
                  oneOf:
                    - type: boolean
                    - type: string
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id:
                    type: integer
                  status:
                    type: string
                    enum: [pending, approved, rejected]
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /verify/status:
    get:
      summary: Verification status
      operationId: getVerificationStatus
      responses:
        '200':
          description: Current verification state
          content:
            application/json:
              schema:
                type: object
                required: [verification_status, latest_document]
                properties:
                  verification_status:
                    type: string
                    enum: [unverified, pending, approved, rejected]
                  latest_document:
                    oneOf:
                      - type: 'null'
                      - $ref: '#/components/schemas/VisaDocument'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /verify/doc/{document_id}:
    get:
      summary: Download verification document (admin)
      operationId: downloadVerificationDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Binary file response
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /admin/verify/pending:
    get:
      summary: List pending verification documents (admin)
      operationId: listPendingDocuments
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Pending documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPendingVerifications'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /verify/pending:
    get:
      deprecated: true
      summary: Deprecated legacy pending-verification alias
      operationId: listPendingDocumentsLegacy
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Pending documents with a deprecation warning header
          headers:
            Warning:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPendingVerifications'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /admin/verify/{document_id}/approve:
    post:
      summary: Approve verification document (admin)
      operationId: approveDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationDecisionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /admin/verify/{document_id}/reject:
    post:
      summary: Reject verification document (admin)
      operationId: rejectDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                review_note:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/VerificationDecisionResponse'
                  - type: object
                    properties:
                      review_note:
                        type: string
                        nullable: true
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /verify/{document_id}/approve:
    post:
      deprecated: true
      summary: Deprecated legacy approval alias
      operationId: approveDocumentLegacy
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationDecisionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /verify/{document_id}/reject:
    post:
      deprecated: true
      summary: Deprecated legacy rejection alias
      operationId: rejectDocumentLegacy
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                review_note:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationDecisionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DocumentId:
      in: path
      name: document_id
      required: true
      schema:
        type: integer
  responses:
    ErrorResponse:
      description: JSON error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error, detail, request_id]
      properties:
        error:
          type: string
        detail:
          type: string
        request_id:
          type: string
...
    VisaDocument:
      type: object
      required: [id, user_id, filename, file_path, file_type, status, waiver_acknowledged]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        filename:
          type: string
        file_path:
          type: string
        file_type:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewer_id:
          type: integer
          nullable: true
        review_note:
          type: string
          nullable: true
        waiver_acknowledged:
          type: boolean
        created_at:
          type: string
          format: date-time
          nullable: true
    PendingVerificationDocument:
      type: object
      required: [id, user_id, filename, created_at]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        filename:
          type: string
        created_at:
          type: string
          format: date-time
          nullable: true
    PaginatedPendingVerifications:
      type: object
      required: [results, count, pagination]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/PendingVerificationDocument'
        count:
          type: integer
        pagination:
          type: object
          required: [page, per_page, total, total_pages, has_next, has_prev]
          properties:
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
            has_next:
              type: boolean
            has_prev:
              type: boolean
...
    VerificationDecisionResponse:
      type: object
      required: [id, status, verification_status]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [approved, rejected]
        verification_status:
          type: string
          enum: [approved, rejected]